+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Replace `--verbose` and `--quiet` with `--log-level`/`-l` and `--no-progress`/`-n`
 - Add support for customizing the number of API retries with `--api-max-retries`
 - Update base bounds script
+- Update filter effectiveness script
 
 ## Nice v3.2.11